+++ release/changelog.md
@@ -1,8 +1,8 @@
 # 🚀 Improvements
-- 
+- Add setup-maven action inputs to configure a custom Maven repository for downloading Develocity and CCUD extensions
 
 # 🐛 Bug fixes
 -
 
 # 📦 Dependency updates
-- 
+-